+++ tests/CircularBuffer/CircularBufferEmptynessTest.c
@@ -3,7 +3,7 @@
 
 
 
-uint8_t circular_buffer_size;
+uint16_t circular_buffer_size;
 CircularBuffer_t buffer_p;
